De-couple Minitron pruning importance estimator from Dynamic Module
