+++ tests/gpu/torch/prune/plugins/test_mcore_mamba_minitron_pruning.py
@@ -23,10 +23,91 @@ skip_if_no_megatron(apex_or_te_required=True, mamba_required=True)
 
 from _test_utils.torch.distributed.utils import spawn_multiprocess_job
 from _test_utils.torch.megatron.models import get_mcore_mamba_hybrid_model
-from _test_utils.torch.megatron.utils import run_mcore_inference_with_dummy_input
+from _test_utils.torch.megatron.utils import (
+    run_mcore_inference,
+    run_mcore_inference_with_dummy_input,
+)
+from _test_utils.torch.misc import compare_outputs, set_seed
 from megatron.core.ssm.mamba_layer import MambaLayer
+from megatron.core.transformer.identity_op import IdentityOp
 
+import modelopt.torch.nas as mtn
 import modelopt.torch.prune as mtp
+from modelopt.torch.prune.plugins.mcore_minitron import (
+    ImportanceEstimatorRegistry,
+    _convert_model_to_dynamic_space,
+)
+
+SEED = 1234
+
+
+def _test_mcore_mamba_parameter_sorting(rank, size):
+    num_layers = size
+    hybrid_override_pattern = "M" * size
+    hidden_size = 256
+    mamba_state_dim = 64
+    mamba_head_dim = 16
+    mamba_num_groups = 2
+    max_sequence_length = 32
+    vocab_size = 64
+    batch_size = 2
+
+    model = get_mcore_mamba_hybrid_model(
+        tensor_model_parallel_size=1,
+        pipeline_model_parallel_size=size,
+        initialize_megatron=True,
+        num_layers=num_layers,
+        hybrid_override_pattern=hybrid_override_pattern,
+        hidden_size=hidden_size,
+        mamba_state_dim=mamba_state_dim,
+        mamba_head_dim=mamba_head_dim,
+        mamba_num_groups=mamba_num_groups,
+        max_sequence_length=max_sequence_length,
+        vocab_size=vocab_size,
+        bf16=False,
+    ).cuda()
+
+    # Randomize norm weights instead of all zeros or ones
+    for n, m in model.named_modules():
+        if "norm" in n and not isinstance(m, IdentityOp):
+            m.weight.data = torch.randn_like(m.weight)
+
+    model.eval()
+    dynamic_space = _convert_model_to_dynamic_space(model)
+    registry = ImportanceEstimatorRegistry(model)  # register imp estimators and forward hooks
+
+    # Compute activations for sorting
+    for _ in range(5):
+        run_mcore_inference_with_dummy_input(model, batch_size)
+
+    # Get the output of the original model
+    prompt_tokens = torch.randint(0, vocab_size, (batch_size, max_sequence_length)).cuda()
+    y1 = run_mcore_inference(model, prompt_tokens)
+
+    mtn.utils.sort_parameters(model)
+    registry.cleanup()
+
+    # check if all mamba_num_heads, mamba_head_dim, hidden_size have been sorted
+    sortable_per_pp = [
+        n for n, hp in dynamic_space.named_hparams(configurable=True) if hp.importance is not None
+    ]
+    # 2 mamba hps per layer + 1 for hidden_size (num_layers is not sorted!)
+    assert len(sortable_per_pp) == 2 * num_layers // size + 1
+
+    # sanity check if the model functionality is preserved after sorting
+    y2 = run_mcore_inference(model, prompt_tokens)
+
+    # check if the inference results after sorting is the same
+    compare_outputs(y1, y2, rtol=1e-5, atol=1e-3)
+
+
+def test_mcore_mamba_parameter_sorting(need_2_gpus):
+    set_seed(SEED)
+    spawn_multiprocess_job(
+        size=torch.cuda.device_count(),
+        job=_test_mcore_mamba_parameter_sorting,
+        backend="nccl",
+    )
 
 
 def _test_mcore_mamba_hybrid_pruning(ckpt_path, rank, size):
