+++ tox.ini
@@ -58,14 +58,11 @@ commands =
 # GPU test environments (Should be used with --current-env)
 ###########################################################
 [testenv:{py310,py311,py312}-cuda12-gpu]
-setenv =
-    MAMBA_FORCE_BUILD=TRUE
 commands_pre =
     # Install deps here so that it gets installed even in --current-env
     pip install -U megatron-core
     pip install git+https://github.com/Dao-AILab/fast-hadamard-transform.git
 
-    # Install Mamba model dependencies (takes 8-10mins!)
     # Skip triton because pytorch-triton is installed in the NGC PyTorch containers
     pip install pip-mark-installed
     pip-mark-installed triton
